% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eif_additive.R
\name{eif_additive}
\alias{eif_additive}
\title{Estimate treatment effects under additivity using the influence function}
\usage{
eif_additive(
  X,
  Y,
  xf = NULL,
  d1_logf = NULL,
  d2_logf = NULL,
  theta_init = NULL,
  calc_se = TRUE,
  ...
)
}
\arguments{
\item{X}{numeric vector, the outcomes of the control observations.}

\item{Y}{numeric vector, the outcomes of the treated observations.}

\item{xf}{numeric vector, points at which the density is given in d1_logf (and d2_logf).
If xf is NULL but d1_logf is supplied, it is assumed that the density was estimated at each points in X.}

\item{d1_logf}{numeric vector, (estimates of) the first derivative of the log density of the control observations X.
If d1_logf is null, the density (and its derivatives) of X is estimated using estimate_density_d_logs.}

\item{d2_logf}{numeric vector, (estimates of) the second derivative of the log density of the control observations X.
Used to estimate standard errors.
If d2_logf is null but calc_se is TRUE (default), the density (and its derivatives) of X is estimated using estimate_density_d_logs.}

\item{theta_init}{numeric, initial value to find the solution to (*). Defaults to the difference in medians.}

\item{calc_se}{logical, should standard errors be calculated? (default = TRUE)
This is slower if d1_logf but not d2_logf is provided as then the density needs to be estimated for standard errors but not for the point estimate.}

\item{...}{additional arguments passed on estimate_density_d_logs for density estimation.}
}
\value{
list of two elements if calc_se = TRUE or one element otherwise: 
        \item{tau}{the point estimate of the treatment effect}
        \item{se}{(if calc_se = TRUE) the estimated standard error}
}
\description{
Efficiently estimate the treatment effect under an additive model based on the influence function.
Finds the parameter \eqn{\theta} such that 
    \deqn{\int f_{0}'(x)/f_{0}(x) f_{0}(x) dx = \int f_{0}'(y-\theta)/f_{0}(y-\theta) f_{1}(y) d y (*)}
and estimates the standard error of \eqn{\hat{\theta}} (under correct specification).
Beware of warnings (about failures to find a solution) coming from pracma::fzero. These are currently not handled properly.
}
\examples{
# draw a random sample with additive treatment effect
X <- rexp(n=1000, rate=2)
Y <- 0.5 + rexp(n=200, rate=2)
eif_additive(X,Y)
}
