% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eif_additive.R
\name{eif_additive_se}
\alias{eif_additive_se}
\title{Estimate the standard error of the efficient estimators under additivity}
\usage{
eif_additive_se(X, p, xf = NULL, d2_logf = NULL, n = NULL, ...)
}
\arguments{
\item{X}{numeric vector, the outcomes of the control observations. (Or a separate sample of observations from the same distribution as the control observations.)}

\item{p}{numeric between 0 and 1, the fraction of the sample that is treated.}

\item{xf}{numeric vector, points at which the density is given in d2_logf.
If xf is NULL but d2_logf is supplied, it is assumed that the density was estimated at each points in X.}

\item{d2_logf}{numeric vector, (estimates of) the second derivative of the log density of the control observations X.
If d2_logf is null, the density (and its derivatives) of X is estimated using estimate_density_d_logs.}

\item{n}{integer, the sample size. Defaults to length(X)/(1-p).}

\item{...}{additional arguments passed on estimate_density_d_logs for density estimation.}
}
\value{
the estimated standard error of the efficient additive treatment effect estimators.
}
\description{
The standard errors can be used for both eif_additive and waq under correct specification of the additive model.
}
\examples{
# draw a random sample of observations
X <- rexp(n=1000, rate=2)
# standard error assuming half of the estimation sample is control
eif_additive_se(X=X,p=0.5)
# standard error assuming there are 2000 control and 200 treated in the estimation sample
eif_additive_se(X=X,p=200/(2000+200),n=2000+200)

}
