% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waq.R
\name{waq}
\alias{waq}
\title{Estimate treatment effects under additivity by weighting quantile differences}
\usage{
waq(X, Y, xf = NULL, d2_logf = NULL, ...)
}
\arguments{
\item{X}{numeric vector, the outcomes of the control observations.}

\item{Y}{numeric vector, the outcomes of the treated observations.}

\item{xf}{numeric vector, points at which the density is given in d2_logf.
If xf is NULL but d2_logf is supplied, it is assumed that the density was estimated at each points in X.}

\item{d2_logf}{numeric vector, (estimates of) the second derivative of the log density of the control observations X.
If d2_logf is null, the density (and its derivatives) of X is estimated using estimate_density_d_logs.}

\item{...}{additional arguments passed on estimate_density_d_logs for density estimation. DO NOT DISABLE ESTIMATION OF DERIVATIVES.}
}
\value{
list of two elements: \item{tau}{the point estimate of the treatment effect} \item{se}{the estimated standard error}
}
\description{
Efficiently estimate the treatment effect under an additive model based on taking a weighted average of the quantile differences:
    \deqn{\sum_{i} w(X_{(i)}) (Y_{(i)} - X_{(i)}) / \sum_{i} w(X_{(i)})}
where \eqn{w(X_{(i)})} is the (estimated) second derivative of the log density of the control, and subscripts (i) indicate that the vectors are sorted.
If X and Y are not of the same length, elements of the shorter vector are duplicated appropriately.
}
\examples{
# draw a random sample with additive treatment effect
X <- rexp(n=1000, rate=2)
Y <- 0.5 + rexp(n=200, rate=2)
waq(X,Y)
}
